requirements.txt
accelerate>=0.19.0
braceexpand>=0.1.7
einops>=0.6.1
einops_exts>=0.0.4
fastapi>=0.95.2
gradio>=3.33.1
huggingface_hub>=0.13.3
importlib_metadata>=6.6.0
inflection>=0.5.1
markdown2>=2.4.8
more_itertools>=9.1.0
nltk>=3.8.1
numpy>=1.23.5
open_clip_torch>=2.16.0
openai>=1.1.1
opencv_python_headless>=4.5.5.64
Pillow>=9.5.0
pycocoevalcap>=1
pycocotools>=2.0.6
Requests>=2.31.0
scipy>=1.10.1
timm>=0.9.2
tqdm>=4.65.0
transformers==4.35.0
uvicorn>=0.22.0
webdataset>=0.2.48
xformers>=0.0.20
natsort>=8.4.0
peft>=0.4.0
litellm>=1.1.1
ijson>=3.2.3
yajl>=0.3.5
deepspeed>=0.10.0
wandb>=0.15.8
trl>=0.5.0
cffi>=1.15.1
pyyaml>=6.0.1
pytest>=7.4.2
prettytable>=3.9.0
